add main page
