+++ src/App.css
@@ -3,6 +3,18 @@
   margin-bottom: 20px;
 }
 
+.main-wrapper{
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin-bottom: 20px;
+  margin-top: 100px;
+}
+
+.button-wrapper{
+  margin-top: 20px;
+}
+
 .investments-wrapper{
   display: flex;
   flex-direction: column;
