+++ src/App.jsx
@@ -4,7 +4,8 @@ import { BrowserRouter, Route, Routes } from 'react-router-dom'
 
 import './App.css';
 import { Investments } from './pages/investments';
-import Main from './pages/main'
+import { Main } from './pages/main';
+import Portfolio from './pages/portfolio'
 
 
 const App = () => {
@@ -12,6 +13,7 @@ const App = () => {
     <BrowserRouter>
       <Routes>
         <Route path="/" Component={Main}/>
+        <Route path="/portfolio" Component={Portfolio}/>
         <Route path="/investments" Component={Investments}/>
       </Routes>
     </BrowserRouter>
